+++ resources/views/Contracts/scripts.blade.php
@@ -1,4 +1,2 @@
-<script type="text/javascript" src="/others/stisla_admin_v1.0.0/dist/js/jquery.smartWizard.min.js"></script>
-<script src="/others/stisla_admin_v1.0.0/dist/modules/nice-select/jquery.nice-select.min.js"></script>
 <script src="/others/stisla_admin_v1.0.0/dist/js/contracts/contract.js"></script>
 <script type="text/javascript" src="/others/stisla_admin_v1.0.0/dist/js/maintenance.js"></script>