+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog Capstone-Laravel-tugboat-dev-PreFinal α
 ### `Pre Final Repository Monday August 27,2018 v.1.0.0`
 
+### Administrator Reports Page Edited
+- [x] Added Changes views for reports
+    - Table View & Chart View
+
+### Consignee Contract Request Page Edited
+- [x] Added request changes modal
+
+### Consignee Contract Request Page Edited
+- [x] Added matrix table and custom matrix suggestion fields
+
+### Contracts Module
+- [x] UI For Contract Page
 &nbsp;
 
 #1.0.2 10-01-2018 α